feat(test): prank proof calldata in tests instead of hardcoding
In all negative tests for proof calldata, we prank the calldata bytes of
a correct proof directly in the test, instead of relying on hardcoded
files. This makes the tests robust between snark verifier changes.

In this commit we temporarily keep all existing failing calldata test
files (for snark verifier v0.1) and assert that our pranking matches the
existing files.

In the next commit we will clean up and remove the hardcoded failing
calldata files.